// src/types/DomainInfo.ts
export interface DomainInfo {
    topic_id: string;
    paymenttransaction_id: string;
    topictransaction_id: string;
    domain: string;
    tld_hash: string;
    sld_hash: string;
    provider: string;
    contract_id: string;
    expiration: number;
    nft_id: number;
    tld: string;
    payer_account_id: string;
    running_hash: string;
    running_hash_version: number;
    sequence_number: number;
    token_id: string;
    network: string;
    nftId: string;
    idx: string;
    account_id: string;
    created_timestamp: string;
    serial_number: number;
    delegating_spender: string;
    deleted: string;
    metadata: string;
    modified_timestamp: string;
}

export interface DomainMetaData {
    transactionId: string;
    nameHash: { domain: string; tld_hash: string; sld_hash: string };
    nftId: string;
    expiration: number;
    provider: string;
    providerData: { contractId: string };
    accountId: string;
}