// src/types/Metadata.ts
export interface MetadataType {
    domain: string;
    eth: string;
    btc: string;
    sol: string;
    avatar: string;
    twitter: string;
    email: string;
    url: string;
    description: string;
    keywords: string;
    discord: string;
    github: string;
    reddit: string;
    telegram: string;
    extras: string;
}